+++ fabric-events-interaction-v0/src/main/resources/fabric-events-interaction-v0.mixins.json
@@ -9,7 +9,6 @@
     "PlayerMixin",
     "ServerGamePacketListenerImplMixin",
     "ServerPlayerGameModeMixin",
-    "ServerGamePacketListenerInteractEntityHandlerMixin",
     "ServerGamePacketListenerImplMixin",
     "ServerPlayerMixin"
   ],
