+++ fabric-game-rule-api-v1/src/client/java/net/fabricmc/fabric/mixin/gamerule/client/RuleListEntryTypeVisitorMixin.java
@@ -68,7 +68,7 @@ public abstract class RuleListEntryTypeVisitorMixin implements GameRuleTypeVisit
 	/**
 	 * @reason We need to display an enum rule's default value as translated.
 	 */
-	@WrapOperation(method = "addEntry(Lnet/minecraft/world/level/gamerules/GameRule;Lnet/minecraft/client/gui/screens/worldselection/EditGameRulesScreen$EntryFactory;)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/gamerules/GameRule;serialize(Ljava/lang/Object;)Ljava/lang/String;"))
+	@WrapOperation(method = "Lnet/minecraft/client/gui/screens/worldselection/AbstractGameRulesScreen$RuleList$1;addEntry(Lnet/minecraft/world/level/gamerules/GameRule;Lnet/minecraft/client/gui/screens/worldselection/AbstractGameRulesScreen$EntryFactory;)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/gamerules/GameRule;serialize(Ljava/lang/Object;)Ljava/lang/String;"))
 	private <T> String displayProperEnumName(GameRule<T> instance, T value, Operation<String> original) {
 		String valueName = original.call(instance, value);
