+++ fabric-game-rule-api-v1/src/client/resources/fabric-game-rule-api-v1.client.mixins.json
@@ -5,7 +5,6 @@
   "client": [
     "AbstractGameRulesScreenAccessor",
     "AbstractGameRulesScreenRuleListAccessor",
-    "AbstractGameRulesScreenRuleListMixin",
     "RuleListEntryTypeVisitorMixin"
   ],
   "injectors": {
