+++ fabric-item-api-v1/src/main/resources/fabric-item-api-v1.accesswidener
@@ -1,2 +1,3 @@
 accessWidener	v2	official
 accessible class net/minecraft/resources/RegistryLoadTask$PendingRegistration
+accessible class net/minecraft/core/component/DataComponentInitializers$BakedEntry
