+++ fabric-resource-conditions-api-v1/src/main/resources/fabric-resource-conditions-api-v1.accesswidener
@@ -1,5 +1,4 @@
 accessWidener	v2	official
 accessible	class	net/minecraft/resources/RegistryOps$HolderLookupAdapter
-accessible class net/minecraft/resources/RegistryDataLoader$ResourceManagerRegistryLoadTask
 accessible class net/minecraft/resources/RegistryLoadTask$PendingRegistration
 accessible field net/minecraft/resources/RegistryOps lookupProvider Lnet/minecraft/resources/RegistryOps$RegistryInfoLookup;
