+++ fabric-resource-conditions-api-v1/src/main/resources/fabric-resource-conditions-api-v1.mixins.json
@@ -6,8 +6,8 @@
     "ReloadableServerResourcesMixin",
     "DataProviderMixin",
     "SimpleJsonResourceReloadListenerMixin",
-    "RegistryDataLoaderPendingRegistrationMixin",
-    "RegistryDataLoaderResourceManagerRegistryLoadTaskMixin",
+    "RegistryLoadTaskPendingRegistrationMixin",
+    "ResourceManagerRegistryLoadTaskMixin",
     "RegistryOpsAccessor",
     "FileToIdConverterAccessor",
     "PackMixin"
